Gherkin tests with projects are green

+++ uservice-webapi/main/host/src/main/java/sinnet/gql/ServiceModel.java
@@ -13,7 +13,7 @@ import sinnet.grpc.common.UserToken;
 public class ServiceModel {
   private @NonNull @Id String projectId;
   private @NonNull @Id String entityId;
-  private @NonNull int entityVersion;
+  private @NonNull long entityVersion;
   private String servicemanEmail;
   private String servicemanName;
   private @NonNull LocalDate whenProvided;
